pyproject.toml
[project]
name = "gdrive-dl-pdf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "docx2pdf>=0.1.8",
    "google-api-python-client>=2.185.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "packaging>=25.0",
    "unoserver>=3.4",
]
